Extracted resources to separate file
